+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/InputDescriptor.kt
@@ -1,12 +1,14 @@
 package at.asitplus.dif
 
+import at.asitplus.dif.rqes.Base64URLTransactionDataSerializer
+import at.asitplus.dif.rqes.TransactionDataEntry
 import com.benasher44.uuid.uuid4
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 /**
  * Data class for
- * [DIF Presentation Exchange v1.0.0](https://identity.foundation/presentation-exchange/spec/v1.0.0/#presentation-definition)
+ * [DIF Presentation Exchange v2.1.1](https://identity.foundation/presentation-exchange/spec/v2.1.1/#term:presentation-definition)
  */
 @Serializable
 data class InputDescriptor(
@@ -20,15 +22,17 @@ data class InputDescriptor(
     val purpose: String? = null,
     @SerialName("format")
     val format: FormatHolder? = null,
-    @SerialName("schema")
-    val schema: Collection<SchemaReference>? = null,
+    /**
+     * Transaction Data is REQUIRED in LSP UC5
+     */
+    @SerialName("transaction_data")
+    val transactionData: List<@Serializable(Base64URLTransactionDataSerializer::class) TransactionDataEntry>? = null,
     @SerialName("constraints")
     val constraints: Constraint? = null,
 ) {
-    constructor(name: String, schema: SchemaReference, constraints: Constraint? = null) : this(
+    constructor(name: String, constraints: Constraint? = null) : this(
         id = uuid4().toString(),
         name = name,
-        schema = listOf(schema),
         constraints = constraints,
     )
 }