+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/AuthenticationRequest.kt
@@ -6,6 +6,7 @@ package at.asitplus.wallet.lib.oidc
 import at.asitplus.catching
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.signum.indispensable.josef.JwsSigned
+import at.asitplus.dif.rqes.UrlSerializer
 import io.ktor.http.*
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
@@ -62,19 +63,6 @@ internal object JwsSignedSerializer : KSerializer<JwsSigned> {
 
     override fun serialize(encoder: Encoder, value: JwsSigned) {
         encoder.encodeString(value.serialize())
-    }
-
-}
-
-internal object UrlSerializer : KSerializer<Url> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("UrlSerializer", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): Url = Url(decoder.decodeString())
-
-    override fun serialize(encoder: Encoder, value: Url) {
-        encoder.encodeString(value.toString())
     }
 
 }